Fix unused variable $machine and replaced with $name parameter
removed $machine.resourcename which doesn't exist in the function context and can result in full clustered resource removal. Replaced with $Name which will evaluate to the machine name that is passed to it. Without using $Name I experienced all my clustered VM's removed instead of just lab ones! 

It may be worth adding additional checks for the removal to ensure only lab resources are going to be targeted.